fix(bin): on linux, resolve the potential symlink to the bash script
On linux, if we create a symbolic link to the binary, the $0 will
report the path in the original context (before resolving the link).
We need to resolve the link to the actual file location to compute
proper working directory for node.

This is useful when people create a symlink from ~/.local/bin/eask ->
install_dir/bin/eask instead of modifying the $PATH.